Refactor for easier testability
