+++ pdf/src/renderer/index.js
@@ -1,62 +1,18 @@
-import FontStore from '@react-pdf/font'
-import renderPDF from '@react-pdf/render'
-import PDFDocument from '@react-pdf/pdfkit'
-import layoutDocument from '@react-pdf/layout'
-import dayjs from '../../common/helpers/dayjs.js'
-import { nodeOps } from './nodeOps.js'
-// eslint-disable-next-line vue/prefer-import-from-vue
-import { createRenderer } from '@vue/runtime-core'
-
-const fontStore = new FontStore()
+import { renderVueToPdfStructure } from './vueRenderer.js'
+import { renderPdfStructureToReactPdf, Font } from './reactPdfRenderer.js'
 
 const pdf = (root, props) => {
-  // For react-pdf, we need an object which will describe the structure of our
-  // pdf document. Our nodeOps will read the Vue component tree and convert it to
-  // the react-pdf-specific data structure inside doc.
-  // For Vue, we need a "root container" (normally the <div id="app"> DOM element).
-  // Vue uses this to keep track of which running Vue app this is.
-  const container = {}
-
   const render = async (compress = true) => {
     // STEP 1: Use Vue primitives (along with our custom logic in nodeOps) to render the Vue
     // component tree into a data structure which react-pdf can understand.
+    const pdfStructure = renderVueToPdfStructure(root, props)
 
-    const { createApp } = createRenderer(nodeOps)
-    const app = createApp(root, props)
-    app.use(
-      {
-        install(app, options) {
-          app.config.globalProperties.api = options.store
-          app.config.globalProperties.$tc = options.$tc
-          app.config.globalProperties.$date = dayjs
-        },
-      },
-      props
-    )
-    app.mount(container)
-
-    // Use the following log statement in case you need to debug the react-pdf object
-    // structure generated by the renderer:
-    //console.log('doc', container.doc)
+    // OPTIONAL: Use the following log statement in case you need to debug the react-pdf
+    // object structure generated by the renderer:
+    //console.log('doc', pdfStructure)
 
     // STEP 2: Pass the generated data structure object to react-pdf.
-
-    const layout = await layoutDocument(container.doc, fontStore)
-
-    const documentProps = container.doc.props || {}
-    const { pdfVersion, language, pageLayout, pageMode } = documentProps
-
-    const ctx = new PDFDocument({
-      compress,
-      pdfVersion,
-      lang: language,
-      displayTitle: true,
-      autoFirstPage: false,
-      pageLayout,
-      pageMode,
-    })
-
-    return renderPDF(ctx, layout)
+    return renderPdfStructureToReactPdf(pdfStructure, compress)
   }
 
   const toBlob = async () => {
@@ -84,10 +40,4 @@ const pdf = (root, props) => {
   }
 }
 
-const Font = fontStore
-
-const StyleSheet = {
-  create: (s) => s,
-}
-
-export { Font, StyleSheet, pdf, createRenderer }
+export { Font, pdf }
